Add initial man page for jfmt tool
There are not many examples yet: that will have to come after the tool
is more fully developed.

I am unclear about the -I option description but that cannot be resolved
until discussion is had so the option remains the same (maybe slightly
different) as the documentation and the tool (though there will be a
slight update to that too).

More has to be done later on, naturally.