Fix bug in vjson_fprint for number conv checks
The function used the wrong macro so that before printing the number it
checked if only converted was true (so parsed is false) when it should
check for if both are true. The other condition in the else if needs to
(and did and still does) check for if converted is false and parsed is
true meaning the number is valid but cannot be represented in a C type.

This was discovered by issue #785 in part but also because the file used
as an example did not trigger a problem with jval or jnamval.

This does not resolve the issue noted, #785, but it at least will allow
the file to be used in the way it needs to be. The issue is an
enhancement as much as it is a bug.