Make !VALID_JSON_NODE condition not an error
In the parse_json_foo() functions it is NOT an error if both converted
and parsed are false. This was brought up on GitHub numerous times (in
issue #752) so I have made this change. I had originally updated it to
use the macro that I added (one of a number of them, see below) as it
previously was an error if !converted but given the purpose of issue #752
and what's been discussed in issue #785 it appears that that check
should not be there at all.

Given that the macros are helpful I will recite (yes I am hijacking the
definition :-) ) them:

    #define VALID_JSON_NODE(item) ((item)->converted || (item)->parsed)
    #define CONVERTED_PARSED_JSON_NODE(item) ((item)->converted && (item)->parsed)
    #define CONVERTED_JSON_NODE(item) ((item)->converted && !(item)->parsed)
    #define PARSED_JSON_NODE(item) ((item)->parsed && !(item)->converted)

Thus VALID_JSON_NODE reports where either of the booleans converted or
parsed is true; CONVERTED_PARSED_JSON_NODE reports that both are true;
CONVERTED_JSON_NODE reports that converted is true but parsed is not;
and PARSED_JSON_NODE reports that parsed is true but converted is false.

New version of json parser and parser: 1.1.1 2023-07-23.