Improve parsed,converted flags in debug output
As parsed and converted booleans are not strictly part of the numbers or
strings but rather just booleans indicating if it's validly parsed and
converted these boolean flags ('p' and 'c') now are suffixed with a ':'.
If both are true then it will look like 'p,c:' but if just parsed is
true, for example, it will look like 'p:'.

Originally I had the thought of comma separating all flags but this
would be problematic as for each one it would have to check if there is
another flag to print and that would greatly complicate things as there
are a lot of flags to inspect.