jval and jnamval -S and -n options now a list
As the options can be specified more than once we need a list of the
options parsed. I am still not clear on when it's OR or AND but that can
be figured out later.

Unfortunately the -n mode does not work correctly in that checking for
numbers being converted seems to be broken, seemingly from a change in
the convenience macros. I will have to address this another time. The -S
option is tested meaning added to list and freeing the list works fine.
It is possible that I made a mistake in the -n processing even though
they're mostly the same.

Some details might change later on but with this I believe all options
are parsed completely. This of course could also change but I believe at
this time it's done.