Fix -n parsing of jval and jnamval
The parsing of numbers now works. This involved a fix in jnamval in the
option -n and a macro that was recently removed added back but with a
simpler definition that is more clear by name. That macro is:

    #define CONVERTED_JSON_NODE(item) ((item) != NULL && (item)->converted == true)

This is important because we require to check that parsed is true but
that it could not be converted.

Now both -n and -S are parsed and each can be used more than once like
the original spec specified.