Update jnamval common code to be in json_util
That is code that is common to both jval and jnamval that still had its
own copy in jnamval to json_util.

Some code is the same but has some minor functional differences and that
code has not been moved over. It might be that it can be later on and at
least the match functions can be later on but the parsing maybe not.
This is not really a problem.

jnamval has options that are not in jval and so in addition to moving
the common code to json_util the code specific to jnamval remains and
will always remain in jnamval*.[ch] itself.

Things not thought of notwithstanding I believe that this really finally
does complete command line processing - unless of course the code that
has minor functional differences should be moved to json_util: that's
not certain as it would require a new variable in a number of functions
which might needlessly complicate matters when the code will probably
not need to be updated again. Also a problem would be bits would have to
be changed in order for it to work properly and that's also probably not
useful.

Something I neglected to point out yesterday is that the code in
(jfmt|jval|jnamval)_test.[ch] is duplicated too except that for common
structures it uses those in json_util. This probably is not necessary
but I'm not sure putting it in json_util is the right approach. Perhaps
a new file that runs tests for the three tools can be added but that can
be decided another time. However since additional tests have to be done
for jval and more still for jnamval maybe this is not a good idea
anyway. This is TBD later - probably much later if it's determined at
all (probably not worth it).