Fix read from stdin in jfmt, jval and jnamval
New version of all three tools: 0.0.5 2023-07-26.

The tools can now work with:

    echo '"test"' | ./jfmt
    echo '"test"' | ./jval
    echo '"test"' | ./jnamval

and also:

    ./jfmt -
    ./jval -
    ./jnamval -

This uses parse_json() instead of parse_json_stream(). We already have
the contents of the stream so this is okay.

Add to json_util struct a char *, json_file_path, which is the file path
specified at the command line.

Check that file path is not empty.