Fixes when filename is NULL in parse_json_stream()
New jparse and json parser version - 1.1.3 2023-07-26.

Now the check if filename == NULL in parse_json_stream() is further up
in the function. The purpose was to if NULL make it read from stdin. The
problem is that if filename is NULL the odds are that stream is also
NULL which means that the making it stdin would never occur.

Also do not call fd_is_ready() if stream is stdin. Changing this fixes a
problem where in some cases the stream would not be read from if stdin.