Update json_util_README.md
Fix header formatting for the three tools.

Some typos were fixed.

Add to common exit codes codes that are in all tools.

Add to common options options that are in all tools.

Added missing exit code to jval.c: no matches was missing.

Added a json file that is mentioned in the json_util_README.md document
for jfmt but with a minor change which was also updated in the
json_util_README.md document as well. The change was changing
'one_tooth' with a value of 'ree' to 'hair' with the value of 'red'.
This is because 'ree' is not a word and I have a weakness for redheads
and besides that it sounds better than "one_tooth":"red" :-)

Sequenced exit codes.

I have not finished the json_util_README.md document for jval and I've
not finished it for jnamval either. That might come later today or might
not be until tomorrow and possibly the next day but at the latest it
should be finished by the end of the week (where Sunday is the last day
of the week: real weeks start with Monday :-) ).