fprintf() -> fpr() and jval/jnamval -c/-C fixes
In jfmt, jval and jnamval when printing the file contents (which is done
in temporary way) use fpr() not fprintf() now that the tty issue is
resolved.

If jval -c/-C or jnamval -c/-C is used without an arg is a command line
error.  This is done in the sanity check function of each tool.