Add -F format to jfmt, jval and jnamval
New version for all tools 0.0.9 2023-07-31.

The optarg 'color' is aliased to 'colour'. Other options are 'default'
and 'pedantic'. Debug output (with DBG_NONE' which has to change - later
on) is added based on the optarg.

The option has NOT been added to the man pages. This will come later
either later today or another day. Most likely another day.

Add missing setting of variables in alloc_jnamval().