Move parse_verbosity(3) to dbg facility
Since the debug facility is the one that uses a verbosity_level that I
added here I have moved the function from jparse/util.c to dbg/dbg.c.

The dbg.3 man page has been updated with this and also some other
improvements. There is a recursive reference to the man page for
parse_verbosity(3) in the man page just for fun.

The dbg related changes have already been added to the dbg(3) repo as a
pull request.

The dbg_example.c file that I wrote for this repo and the dbg(3)
repo has also been updated with a comment change, referring to not
strtol(3) but rather the parse_verbosity(3) function.

The dyn_array/dyn_test.c no longer has the parse_verbosity() function as
this is in the dbg(3) facility which the dyn_array facility uses.