Fix dyn_test.c wrt parse_verbosity(3)
To make the dyn_test.c work in the dyn_array repo a new define has to be
added and checked so that if the dbg repo is not in use strictly (as in
not in this repo - see below) then it provides the parse_verbosity()
function itself.

Perhaps the macro should be MKIOCCCENTRY_USE like the other place this
is used but currently it is DBG_USE instead.