Add -I to jnamval (same as from jfmt)
New version of jnamval is 0.0.11 2023-08-05.

Add -I option with the same semantics as jfmt. This option does not
exist in jval and it is not clear if it is supposed to. It would be very
easy to add as it's common code but it might not be necessary. This is
TBD later.

Updated jnamval.1 to add the -I option.

Fix typo in jnamval usage string that could cause confusion - it
referred to a non-existent option. It was supposed to be an option arg
and it now is correct.

Typos were discovered in jfmt.1, jval.1 and jnamval.1 when updating
jnamval.1 and these were also fixed.