Fix jnamval -t and -P lists
The types list allowed by -t has been fixed. As it's now the same in
both jval and jnamval it's in json_util.[ch] not in the files for both
jval and jnamval.

There was a fix in one of the parse functions where it did not allow for
a comma separated list but only one item which was incorrect.

Many name changes in macros and function names. An unfortunate
complication with the jnamval -p and -P options is that since both are
for print types there is a possible confusion with names of macros as
there are two sets of JNAMVAL_PRINT_ macros but whether one of the sets
should be renamed is not yet determined. If they are the function names
might also want to change as both -p and -P have similar function names
for checking what options are used. Previously the -p option referred to
matching of name, value, both or json but it has to do with printing,
not matching. The -t option is matching types. Maybe the -p needs to be
changed to a different letter but this can be determined in discussion.

Updated jnamval.1 man page.