cleanup .gitignore
Added `dbg/.gitignore`, `dyn_array/.gitignore`, and `jparse/.gitignore`
as these are the main directories of trees that are "cloned" from
other GitHub repos.  Those other GitHub repos need their own main
directory ".gitignore" file and so we add these new files in
anticipation that they will use them.

Entries that belong to `dbg/.gitignore`, `dyn_array/.gitignore`,
and `jparse/.gitignore` were removed from the top level `.gitignore`
file.

We removed many of the "generic exclusion lines" from the top level
`.gitignore` file.  A "generic exclusion line" refers to an exclusion
that is anywhere in the tree.  Normally "generic exclusion lines"
are the business of a given user's own "~/.gitignore" file.

We did leave behind 5  "generic exclusion lines" in the above
mentioned .gitignore files:

    *.[ao]
    *.dSYM/
    *~
    .*.swp
    .DS_Store

It can be argued that even these files do not belong in a GitHub
.gitignore files.  We kept these 5 around as a compromise and
to assist users who do not have a well setup  "~/.gitignore" file.

We made all of the non-generic exclusion lines "/-path" based.

A number old entries in the `.gitignore` file that have no effect
were removed.  It is possible that we removed something that is
still needed for certain more obscure cases.  If that is true, then
a later update to a .gitignore file can correct such a mistake.