release dbg version 3.0 2023-08-05
The dbg version was changed from "2.10 2023-08-01" to "3.0 2023-08-05"
because the function `parse_verbosity()` was added to the interface.

The `parse_verbosity()` was simplified to only take a "optarg" string.
It longer exits and thus does NOT need to know the name of the
calling program (or function).

A new DBG symbol, DBG_INVALID, is returned by `parse_verbosity()`
if an error is detected.

The return from `parse_verbosity()` should checked for values < 0
by the calling function.

Updated the `dbg(1)` man page according to the above.