release dyn_test version 1.10 2023-08-05
Changed DYN_TEST_VERSION from "1.9 2023-02-04" to "1.10 2023-08-05".

Changes to dyn_test are in responce to release dbg version 3.0
2023-08-05.  The dbg version was changed from "2.10 2023-08-01" to
"3.0 2023-08-05" because the function `parse_verbosity()` was added
to the interface.

The `dyn_test` uses an static internal copy of `parse_verbosity()`,
unless it is compiled with DBG_USE defined, in which case the
`parse_verbosity()` function is taken from `dbg.a`.  By default,
`dyn_test` is made with DBG_USE defined and thus by default `dyn_test`
uses the `parse_verbosity()` function from `dbg.a`.

Only the test code was impacted.  The `dyn_array.a` library
does not use `parse_verbosity()`.