Add struct json *node to some struct json_ structs
The following struct json_foo structs were modified to have a struct
json *node so one can make use of the parse functions of a type of
struct json_foo for explicit parsing of a JSON node outside of the
parsing of a document:

    json_number
    json_string
    json_boolean
    json_null

The idea of parsing a specific JSON node outside of a document is for
the jval and jnamval options -S and -n which have to verify that the
values specified in the options are valid JSON as the comparison will
need the values in the structs json_string and json_number depending on
the options of the tools. This feature might be useful in other cases as
well but at this point only jval and jnamval use it.

The function json_alloc() explicitly sets node to NULL and the
parse_json_(number|string|boolean|null) set the item->node to the
respective struct json *node.

It might be that other json_ structs need to be modified but this is TBD
later. If so the functions json_alloc() and the respective parse_json_
will have to be updated in a similar manner as the above mentioned ones.

The tool jnum_gen has been updated to explicitly set the node to NULL in
the test_result array. Updated jnum_test.c through make rule
rebuild_jnum_test.

This member to the structs has not been added to the man page. Whether
this should be done depends on how the man page (jparse.3) is
constructed but if it does not explicitly list the struct json_
structures it might be good to add them. This would also be good to add
to the README.md file.

Fix make picky for jnamval_util.h which happened because of the prefix
name change of JNAMVAL_PRINT_ to JNAMVAL_RESTRICT_. This was done by
changing two (though only one line was a problem) macros which are just
some bitwise ORs to be two lines instead of one (the usual \ at the
EOL).

New version of JSON parser and jparse: 1.1.5 2023-08-08.

New version of jval and jnamval: 0.0.14 2023-08-08.

New version of jnum_gen: 1.0.1 2023-08-08.