Add bad manifest filenames tests
Each forbidden filename in the manifest now has a separate file to check
it including prog.alt. The -34.json file was also updated to have
prog.alt as an extra filename though it might be better to also check
this for other files too if this is not done (been too long and too many
other changes in the other repo to recall if this is done or not).

That means that under test_ioccc/test_JSON/info.json/bad individual
files that have an extra file with the names of:

- README.md
- index.html
- inventory.html
- prog
- prog.alt
- prog.orig
- prog.orig.c

have been added. Unless other files have to be added later to the list
of disallowed filenames or there needs to be a more specific check in
mkiocccentry for filenames of the required files this should complete
the disallowed and required filenames checks.