improve location ZZ
Fixed double space in comment for XX.

Changed "ZZ" location from "User-assigned" to "Unknown location".

NOTE: Those authors currently with a "location_code" of JSON null,
and/or those authors currently with a "location_name" of JSON null
will now use a "location_code" of "ZZ" and a "location_name" of
"Unknown location".