add common_name to locations and break things
We must add the notion of a _common_name_ as well as an "official" _name_
to a given location.  The IOCCC code uses the "official" _name_
only (for now?), so one must use `-c` to the `location(1)` tool to
cause it to work with _common_name_'s instead of the default
"official" _name_'s.

This breaks tests.  Sorry (tm Canada 🇨🇦).

TODO: Determine if we really need _location_name_ in JSON files
such as `.auth.json`.  We do **NOT** think we want the
_location_name_ used and that _location_code_ is sufficient.