Improve rebuilding txzchk test error files
The script test_ioccc/test_txzchk.sh now has a new option: -B. This
rebuilds the test error files. The script prompts the user to verify
that everything is correct. If it does not get a YES (case sensitive!)
it will exit 3 (invalid command line). If this is the case the rest of
the rule is not run so no notice about adding and committing the new or
updated files.

The rule in the Makefile now is called (the previous one still exists
though): rebuild_txzchk_test_errors.

The test_ioccc/test_txzchk/README.md file has been updated for this
(there was also an error in it but that text has been removed as the
make rule tells the user what to do which is what the incorrect text was
only of course the text is now corrected).