Improve txzchk_test.sh -B option
The script itself now tells the user what to do when it is done (run the
script and then if all is okay commit the files, assuming that there are
any changes). The Makefile rule (in the top level directory) no longer
does this. The message has also been revised to say that if something is
wrong then [you] will have to fix the problems before proceeding.