Release the full version 1.0.62 2024-05-15
The previous commit was incomplete.

Changed MKIOCCCENTRY_REPO_VERSION from "1.0.1 2023-02-14"
to "1.0.62 2024-05-15" to match the above release version.

Changed IOCCC_CONTEST from "IOCCCMOCK"
to "IOCCC28" for testing purposes.  :-)

Changed IOCCC_YEAR from "2023"
to "2024" for testing purposes.  :-)

Changed MKIOCCCENTRY_VERSION from "1.0.4 2024-03-04"
to "1.0.5 2024-05-15" because of the above constant changes.

The filename formed by _mkiocccentry(1)_ was changed
to start with _submit._ instead of _entry._.

Changed FNAMCHK_VERSION from "1.0.1 2024-03-02"
to "1.0.2 2024-05-15" because of the above change.

Updated `mkiocccentry(1)` and `chkentry(1)` man pages accordingly.

Updated the test files according to the above changes.