Make txzchk -q not show output
This only happened when parsing a tarball (real tarball, not a test text
file output of tar). It still exits 0 if no issues are found.

I have not updated the version of the tool as I do not at this time have
the time to update the test files but that can happen later.