Add some credits to location code
To match what was done in chkentry I have also done it in the location
tool code in the same way that was done in the chkentry code and header
files.

Also in CHANGES.md I removed a reference to '@xexyl' (a fun pun is
attributed to me still, however). This is because otherwise every time a
release is made GitHub would tell me I was mentioned which is both
confusing (as it's buried deeply) and unnecessary (plus if ever I am
mentioned I might otherwise miss it). This problem is actually how I
discovered the missing credits mentioned in this commit.